Micom Module - improve angular rotation motion   -> compensate angulara velocity for rotation motion
- handling rapid change of rotation direction

reset lastTheta value when Reset() is called.

Change function in IMU sensordevice
- GetOrientation(): Quaternion -> euler angle

+++ Assets/Scripts/Devices/MicomInput.cs
@@ -77,7 +77,6 @@ public class MicomInput : Device
 
 			// Set reversed value due to differnt direction
 			// Right-handed -> Left-handed direction of rotation
-
 			if (micomWritingData.Value.IntValue == 0)
 			{
 				controlType = VelocityType.LinearAndAngular;
@@ -87,6 +86,8 @@ public class MicomInput : Device
 
 				angularVelocity
 					= (!child1.Name.Equals("AngularVelocity")) ? 0 : (float)-child1.Value.DoubleValue;
+
+				// Debug.Log(linearVelocity.ToString("F10") + ", " + (-child0.Value.DoubleValue).ToString("F10") + " | " + angularVelocity.ToString("F10") + ", " + (-child1.Value.DoubleValue).ToString("F10"));
 			}
 			else if (micomWritingData.Value.IntValue == 1)
 			{
@@ -140,21 +141,21 @@ public class MicomInput : Device
 
 	public float GetWheelLeftVelocity()
 	{
-		return (float)wheelLinearVelocityLeft;
+		return wheelLinearVelocityLeft;
 	}
 
 	public float GetWheelRightVelocity()
 	{
-		return (float)wheelLinearVelocityRight;
+		return wheelLinearVelocityRight;
 	}
 
 	public float GetLinearVelocity()
 	{
-		return (float)linearVelocity;
+		return linearVelocity;
 	}
 
 	public float GetAngularVelocity()
 	{
-		return (float)angularVelocity;
+		return angularVelocity;
 	}
 }
